Rename StudyParticipant -> StudyParticipation
This better suits the nature of this cross reference table and makes the work of enhancing studies (in another PR) somewhat easier.
Also rename participant_id -> patient_id on study_participations to make the relationship clearer.

Note the tigger on study_participations that generates a pseudo-encrypted external_id will
continue working as Rails kindly re-wires it during the migration. The name of the trigger and the
function it calls still use the name 'participant' but we can live with this for now.