Allow user authentication via LDAP
If Renalware.config.ldap_authentication = true for an environment, authenticate users only if they exist in the directory service
specificed in .env with LDAP_* env variables and having a matching password, and they have an entry in the users table in Renalware.
This is an initial commit and there remain some things to do. See PR.