Check for current user before checking role when displaying error details
Currently when we display a 500 error, we check if the user is a superadmin and if they are we display the full error message to them. However if a user is not logged in and there is an error, the original error if obfuscated by the error generated in current_user_is_super_admin? which fails if current_user is nil.
This change checks for a nil current_user