HL7 clinic mapping
If the option to use "by_name_mapping" is set, we will try to resolve
the clinic by looking up the HL7 clinic name in a clinic_mappings table.
If no mapping exists, we create one with the default clinic (if any) and
return that clinic. If no default clinic exists, we create the mapping
with a nil clinic_id and return nil.
This lets us map a specific clinic (sometimes including consultant) eg
Nephrology New Steve Smith to a more generalised clinic eg
Nephrology New in Renalware.