Remove Patient::Abridgement and the master_patient_index table
The concept of the master patient index was introduced but never used - it was to help a user when manually adding a patient, by matching (or suggesting a near match) from the master patient index (build from HL7 data), and copying in the found demographics from there.
It was never used, and as it slightly moot now since that, if you are receiving HL7 outpatient and referrals which themselves add the patient, there should in theory be limited need to add patients manually (and therefore need the master patient index in theory). It will still always be possible to add a patient manually (even if limited to superadmins) but I can't see us doing the work of searching/match against the master patient index at this stage.

The table has been renamed to patient_master_index_deprecated rather than removed, just in case someone is using it.