Copy HD session summary as text to clipboard
Required for Imperial to allow a user to paste the summary information
for a complete HD Session into their EPR system.
Display a clipboard button/link in each patient HD session row, and also
on the HD session show page.
When clicked, this sends an AJAX request to a new
session_clipboards#show action which renders a text template
suitable for pasting into their EPR system. The stimulus controller
then copies this text to the clipboard and shows a tooltip to confirm
success. The stimulus controller falls back to using a textarea
if the clipboard API is not available (e.g. in older browsers).